pyproject.toml
[project]
name = "local-camera-view"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fastexcel>=0.14.0",
    "opencv-python>=4.12.0.88",
    "polars>=1.32.3",
    "python-dotenv>=1.1.1",
    "qrcode>=7.4.2",
    "uvicorn>=0.35.0",
    "pyffmpeg>=2.0.0",
    "numpy>=1.24.0",
    "asyncio-mqtt>=0.16.0",
    "aiohttp>=3.9.0",
    "pillow>=11.3.0",
]
